clean_requirements.txt
accelerate==1.1.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.4.0
appdirs==1.4.4
astor==0.8.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Authlib==1.3.1
baron==0.10.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
boto3==1.35.43
botocore==1.35.43
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
clamd==1.0.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.5
datasets==3.1.0
dill==0.3.8
distro==1.8.0
dnspython==2.6.1
docker==7.1.0
docstring_parser==0.16
email_validator==2.2.0
environs==9.5.0
eval_type_backport==0.2.0
ewmh==0.1.6
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
graphql-core==3.2.3
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-health-checking==1.62.3
grpcio-status==1.62.3
grpcio-tools==1.62.3
gTTS==2.3.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langsmith==0.1.84
lazy_loader==0.4
libcst==1.4.0
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.2
mdurl==0.1.2
milvus-lite==2.4.10
mistralai==1.1.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.35.12
opencv-python==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
playsound==1.2.2
playwright==1.42.0
pluggy==1.5.0
pooch==1.8.2
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyarrow==18.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pyclip==0.7.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==11.0.1
pygame==2.5.2
PyGetWindow==0.0.9
PyJWT==2.9.0
pymilvus==2.4.9
pymongo==4.8.0
PyMsgBox==1.0.9
pyparsing==3.1.2
PyPDF2==3.0.1
pyperclip==1.9.0
pyperclip-plus==0.1.9
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
pytest==8.3.3
pytest-asyncio==0.23.7
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0
RapidFuzz==3.10.1
readability-lxml==0.8.1
redbaron==0.9.2
redis==5.0.7
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.19.0
rply==0.7.8
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
sentence-transformers==3.0.1
shapely==2.0.6
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.5.0.post1
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
strawberry-graphql==0.235.2
sympy==1.13.0
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.3.3
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.2
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.46.2
triton==2.3.1
tweepy==4.14.0
typer==0.12.3
typing-inspect==0.9.0
tzdata==2024.1
ujson==5.10.0
ultralytics==8.2.58
ultralytics-thop==2.0.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.19.0
validators==0.28.3
watchfiles==0.22.0
weaviate-client==4.6.5
websockets==12.0
xxhash==3.5.0
yarl==1.9.4
